edurag/backend/requirements.txt
# EduRAG Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
openai>=1.0.0
tiktoken>=0.5.0
pgvector>=0.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
gunicorn==21.2.0

# RAG & PDF Processing
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
pdfplumber>=0.11.0
pypdf>=4.0.0

# OCR para PDFs escaneados
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
